numeric domains abstract values as single instances, javadoc. moved domains in analysis package
