+++ lisa/imp-testcases/tutorial/program.imp
@@ -1,0 +1,60 @@
+class phdcourse {
+
+	constants(a) {
+		b = a;
+		c = 1;
+   		while (b < a) 
+        	b = b + c;
+   		return a;
+    }
+    
+	sign_parity_example() {
+		i = 2;
+		max = 10;
+		while (i < max) 
+			i = i + 1;
+		return i;
+	}
+	
+	constants(a) {
+		b = 0;
+		c = 1;
+   		while (b < a) 
+        	b = b + c;
+   		return a;
+    }
+	
+	intv_dec(c,d) {
+		i = 1000;
+		while (i > 0) 
+			i = i - 1;
+			
+		return i;
+	}
+	
+	gcd(a, b) {
+   		while (a != b) 
+        	if (a > b) 
+            	a = a - b;
+        	else
+            	b = b - a;
+   		return a;
+    }
+	
+	div(i,j) {
+		if (j != 0)
+			i = i / j;
+		else
+			i = j / i;
+		return i;
+	}
+
+	ub_example() {
+		if (y < z) 
+			x = y - 1;
+		else
+			x = z - 1;
+			
+		return x;
+	}
+}
